TRANSCEIVER PORT
This update is a breaking change with respect to the clients
`gmt_dos-clients_transceiver/scope/io/core` and the `interface`
The port # for the transceiver is now read directly from the `UID` type
and is not longer passed as argument

+++ clients/core/src/operator.rs
@@ -64,7 +64,7 @@ where
 {
     type DataType = <U as UniqueIdentifier>::DataType;
 
-    const PORT: u32 = <U as UniqueIdentifier>::PORT;
+    const PORT: u16 = <U as UniqueIdentifier>::PORT;
 }
 
 pub struct Right<U: UniqueIdentifier>(PhantomData<U>);
@@ -74,7 +74,7 @@ where
 {
     type DataType = <U as UniqueIdentifier>::DataType;
 
-    const PORT: u32 = <U as UniqueIdentifier>::PORT;
+    const PORT: u16 = <U as UniqueIdentifier>::PORT;
 }
 
 impl<T, U> Read<Left<U>> for Operator<T>
