+++ clients/io/src/gmt_m1.rs
@@ -34,7 +34,7 @@ pub mod assembly {
     impl Assembly for M1RigidBodyMotions {}
     impl UniqueIdentifier for M1RigidBodyMotions {
         type DataType = Vec<f64>;
-        const PORT: u32 = 50_006;
+        const PORT: u16 = 50_006;
     }
 
     /// Hardpoints displacements `[cell,mirror]`
@@ -42,7 +42,7 @@ pub mod assembly {
     impl Assembly for M1HardpointsMotion {}
     impl UniqueIdentifier for M1HardpointsMotion {
         type DataType = Vec<Arc<Vec<f64>>>;
-        const PORT: u32 = 50_007;
+        const PORT: u16 = 50_007;
     }
 
     /// Hardpoints forces
@@ -50,7 +50,7 @@ pub mod assembly {
     impl Assembly for M1HardpointsForces {}
     impl UniqueIdentifier for M1HardpointsForces {
         type DataType = Vec<Arc<Vec<f64>>>;
-        const PORT: u32 = 50_007;
+        const PORT: u16 = 50_007;
     }
 
     /// Actuators command forces
@@ -58,7 +58,7 @@ pub mod assembly {
     impl Assembly for M1ActuatorCommandForces {}
     impl UniqueIdentifier for M1ActuatorCommandForces {
         type DataType = Vec<f64>;
-        const PORT: u32 = 50_008;
+        const PORT: u16 = 50_008;
     }
 
     /// Actuators applied forces
@@ -66,7 +66,7 @@ pub mod assembly {
     impl Assembly for M1ActuatorAppliedForces {}
     impl UniqueIdentifier for M1ActuatorAppliedForces {
         type DataType = Vec<Arc<Vec<f64>>>;
-        const PORT: u32 = 50_008;
+        const PORT: u16 = 50_008;
     }
 }
 
@@ -76,43 +76,43 @@ pub mod segment {
     /// Force andf moment at center of gravity
     pub enum BarycentricForce<const ID: u8> {}
     impl<const ID: u8> UniqueIdentifier for BarycentricForce<ID> {
-        const PORT: u32 = 56_001 + 100 * ID as u32;
+        const PORT: u16 = 56_001 + 100 * ID as u16;
         type DataType = Vec<f64>;
     }
     /// Rigid body motion (Tx,Ty,Tz,Rx,Ry,Rz)
     pub enum RBM<const ID: u8> {}
     impl<const ID: u8> UniqueIdentifier for RBM<ID> {
-        const PORT: u32 = 56_002 + 100 * ID as u32;
+        const PORT: u16 = 56_002 + 100 * ID as u16;
         type DataType = Vec<f64>;
     }
     /// Hardpoints displacements `[cell,mirror]`
     pub enum HardpointsMotion<const ID: u8> {}
     impl<const ID: u8> UniqueIdentifier for HardpointsMotion<ID> {
-        const PORT: u32 = 56_003 + 100 * ID as u32;
+        const PORT: u16 = 56_003 + 100 * ID as u16;
         type DataType = Vec<f64>;
     }
     /// Hardpoints forces
     pub enum HardpointsForces<const ID: u8> {}
     impl<const ID: u8> UniqueIdentifier for HardpointsForces<ID> {
-        const PORT: u32 = 56_004 + 100 * ID as u32;
+        const PORT: u16 = 56_004 + 100 * ID as u16;
         type DataType = Vec<f64>;
     }
     /// Actuators applied forces
     pub enum ActuatorAppliedForces<const ID: u8> {}
     impl<const ID: u8> UniqueIdentifier for ActuatorAppliedForces<ID> {
-        const PORT: u32 = 56_005 + 100 * ID as u32;
+        const PORT: u16 = 56_005 + 100 * ID as u16;
         type DataType = Vec<f64>;
     }
     /// Actuators command forces
     pub enum ActuatorCommandForces<const ID: u8> {}
     impl<const ID: u8> UniqueIdentifier for ActuatorCommandForces<ID> {
-        const PORT: u32 = 56_006 + 100 * ID as u32;
+        const PORT: u16 = 56_006 + 100 * ID as u16;
         type DataType = Vec<f64>;
     }
     /// SEGMENT RBM DOF selector ([0,...,6]->[Tx,Ty,Tz,Rx,Ry,Rz])
     pub enum M1S<const ID: u8, const DOF: u8> {}
     impl<const ID: u8, const DOF: u8> UniqueIdentifier for M1S<ID, DOF> {
-        const PORT: u32 = 56_001 + 10 * (1 + DOF) as u32 + 100 * ID as u32;
+        const PORT: u16 = 56_001 + 10 * (1 + DOF) as u16 + 100 * ID as u16;
         type DataType = Vec<f64>;
     }
 }
